9. Create a DetailView that will accept a value named child (type Child) passed from the ListView. The DetailView should be set up so that it looks like the Image below, with alignment and padding/spacing as shown. Use variables named: firstName, lastName, naughty, smacks, and notes User interface setup includes: - All Text Fields / Text Views / Labels for UI elements below should be in Title 2 font, except for the large number below the Stepper View, which is in Large Title font. - Text Views: o Three Text Views to label text entry areas for "First Name:", "Last Name:", "Notes:" in bold. - Text Fields: o Two Text Fields, with text prompt "first" and "last", respectively, both in Rounded Boarder style (no need to overlay). o A third Text Field for notes with prompt "notes" that can also scroll vertically if more than one line of notes is entered. Also use Rounded Border style and positioned as shown.
- A Toggle View labeled "Naughty?" in Bold.
- Stepper View / Text Field:

5
o labeled "Smacks Deserved:" in Bold, which will allow the user to select smacks in a range
from 0 through 5.
o The number of smacks should be in a Text View centered below, bold, in Large Title font,
and change as the stepper is pressed within the range of acceptable values.

- Image Views:
o Two Image Views, "boy" and "girl" side-by-side, with bottom at the bottom of the window,
sized roughly the same size inside at 250 points tall.
o The images should be side-by-side centered horizontally in the View (as shown).
- Surround all elements in standard padding